Update petfeederProject.ino
PetFeeder is a Wifi-controlled device to feed our pets/animals who are alone at home. After accessing the PetFeeder application, we start the device by clicking the 'Feed' button. The motor inside the device rotates and empties the food into the container as needed. In this way, we can feed our animals both inside and outside the house in a remotely controlled manner.